large-language-models/metadata-enrichment/database-table/requirements.txt
dotenv~=0.9.9
langchain-community==0.3.17
langchain-openai==0.3.5
langgraph==0.2.72
pandas~=2.2.3
psycopg2-binary~=2.9.10
SQLAlchemy~=2.0.39
